Add files via upload
Some of the steps could not be completed as I did not have adequate time to gather resources, some of which were to be bought. I have however included the library used, screenshots of the output attained as well as a schematic drawing of the wiring. For the MQTT server, I used the google extension version and got the error attached as well. I am not as familiar with MQTT and any feedback will be appreciated.